requirements.txt
pymongo==4.6.2
bcrypt==4.1.2
prettytable==3.9.0
apscheduler==3.10.1
pytz==2023.3
python-dotenv==1.0.0